fix(web): show interview topics on meeting outline + person-page bar (#59)
* fix(models): serialize confidence as native float, not numpy scalar

scipy's cosine similarity returns a numpy float, so voice-match
confidence was leaking numpy scalars into the serialized Segment /
SpeakerMapping dicts (and thus JSON/DB payloads). Coerce to native
float at the match site and at both to_dict boundaries.


* fix(web): show topic outline for interviews (topic section_type)

The meeting page built its topic outline by filtering sections against a
SUBSTANTIVE set of {discussion, public_comment, consent_agenda, vote} —
the same set that omitted "topic" on the pipeline side (fixed in #58).
Interview/media summaries emit section_type="topic", so every interview
section was filtered out and the outline (with its topic labels) rendered
empty even though the API returns the tagged sections.

Extract the filter into lib/outline.ts (SUBSTANTIVE_SECTION_TYPES now
includes "topic"), use it from MeetingDetailClient, and add a unit test.


* fix(web): surface meeting topics on person-page bar instead of "untagged"

The person-page meeting bar showed the topic of the section currently
scrolled into view, and "— procedural / untagged" whenever the meeting
was collapsed or scrolled to an untagged intro. For interviews (which
open with an untagged Introduction) this read as "no topics" on load.

Now: when actively scrolling a tagged section, show that one topic (the
sticky indicator, unchanged); otherwise fall back to the meeting's
distinct topics via meetingTopics(). Only genuinely topic-less meetings
show "untagged". Add meetingTopics() + tests to lib/outline.ts.


---------

+++ src/identify.py
@@ -73,7 +73,7 @@ def match_voice_profiles(
                 mappings[label] = SpeakerMapping(
                     speaker_label=label,
                     speaker_name=display,
-                    confidence=round(best_score, 3),
+                    confidence=float(round(best_score, 3)),
                     id_method=method,
                     needs_review=(display is None),
                 )
